[core] Move `CommandAllocator` into its own module.
No intended change in behavior.
